# core/api/rest/openapi.yaml
...
openapi: 3.1.0
info:
  title: CypherPanel API
  description: >-
    Operator-facing control-plane API covering Phases 1–3: authentication and
    teams/roles, the server lifecycle, projects/environments, applications and
    the deploy pipeline, managed databases with S3 backups/restore, preview
    environments, notifications, and scheduled tasks. Authentication is a bearer
    session token from /api/v1/auth/login; the GitHub webhook authenticates by
    per-application HMAC instead. Every project-scoped route is authorized by
    team membership (a non-member sees 404, insufficient rank 403).


    Every response — success, error and SSE stream alike — carries an
    `X-Request-Id` header (`components/headers/RequestId`), and every JSON
    error body repeats it as `trace_id`. It is the value to quote in a bug
    report and the key to search for in `GET /api/v1/panel/logs`. Individual
    responses reference the header only where a generated client benefits;
    it is present on all of them.
  version: 0.3.0
servers:
  - url: /
...
paths:
...
  /api/v1/deployments/{id}:
    get:
      operationId: getDeployment
      tags: [deployments]
      summary: One deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The deployment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: No such deployment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /webhooks/github/{id}:
    post:
      operationId: handleGithubWebhook
      tags: [deployments]
      summary: GitHub push webhook (HMAC-authenticated, no session)
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The application's webhook id (wh_… prefix).
        - name: X-Hub-Signature-256
          in: header
          required: true
          schema:
            type: string
          description: GitHub's HMAC-SHA256 body signature ("sha256=<hex>").
      responses:
        "202":
          description: >-
            Push on the configured branch; deployment started — or, when the
            environment requires approval, parked with
            `status: awaiting_approval` and a `requested_by`-less approval
            (a push has no panel user behind it).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
        "204":
          description: Authenticated, but not a deployable push (other branch, deletion, non-push event).
        "401":
          description: Signature verification failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Unknown webhook id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: >-
            The environment is inside a freeze window; nothing was written. The
            error names the window and when it lifts, so the delivery is
            diagnosable from the response alone and can be redelivered from
            GitHub after the window.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: Session token from /api/v1/auth/login.
  headers:
    RequestId:
      description: >-
        Correlation id for this request, minted by the API's outermost
        middleware and present on every response. It is repeated as `trace_id`
        in every JSON error body, and it is the key to search for in
        `GET /api/v1/panel/logs`. An inbound `X-Request-Id` is honoured only
        when the peer is a configured trusted proxy and the value is a short
        printable token; otherwise a fresh id is minted.
      schema: { type: string }
      example: trace_9f3a-11bd-4c02-7e51
  responses:
...
    Unauthorized:
      description: Missing, invalid, or expired session token.
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: Authenticated but the caller's role is insufficient for this action.
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable message; never contains secrets.
        trace_id:
          type: string
          description: >-
            The value of this response's `X-Request-Id` header, repeated in the
            body so a screenshot of a failure carries it. Paste it into a bug
            report or `GET /api/v1/panel/logs` to find the request. Present on
            every response served through the API's middleware.
          example: trace_9f3a-11bd-4c02-7e51
        retry_after_seconds:
          type: integer
          minimum: 1
          description: >-
            Present only on `429`: whole seconds until the throttle lifts, the
            same value as the `Retry-After` header. Count down against it rather
            than guessing.
...
    Deployment:
      type: object
      required: [id, application_id, revision_id, status, trigger, created_at]
      properties:
        id:
          type: string
        application_id:
          type: string
        revision_id:
          type: string
        status:
          type: string
          description: >-
            Pipeline state. `awaiting_approval` is a deploy that deploy
            protection parked before it started: nothing was published, the
            application's own status is untouched, and it holds no queue slot.
            It is NOT terminal — it is waiting for a person. A rejected deploy
            ends `failed`, with a detail naming the rejecter.
          enum: [queued, building, distributing, rolling_out, succeeded, failed, awaiting_approval]
        trigger:
          type: string
          enum: [manual, webhook, rollback]
        detail:
          type: string
        approval:
          description: >-
            The gate decision on this deployment, present only when it was
            gated. Carries enough to render the pending card without a second
            request; a deployment that was never gated omits it entirely.
          oneOf:
            - $ref: "#/components/schemas/DeployApproval"
            - type: "null"
        created_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
          nullable: true
...
    DeployApproval:
      type: object
      required: [deployment_id, environment_id, required_role, state, created_at]
      properties:
        deployment_id: { type: string }
        environment_id: { type: string }
        requested_by:
          type: string
          description: >-
            The user who asked for the deploy. Empty for a webhook deploy — a
            push has no panel user behind it — and for a requester who has since
            been deleted.
        requested_by_email:
          type: string
          description: That user's address, joined for display. Empty when there is no requester.
        required_role:
          type: string
          enum: [member, admin, owner]
          description: The rank that must approve, snapshotted at park time.
        state:
          type: string
          enum: [pending, approved, rejected]
        decided_by:
          type: string
          description: Who approved or rejected; empty while pending.
        decided_by_email:
          type: string
        decided_at:
          type: string
          format: date-time
          nullable: true
        reason:
          type: string
          description: The rejecter's sentence, stored verbatim and rendered as text. Empty on an approval.
        created_at:
          type: string
          format: date-time
...
security:
  - bearer: []